[IID Cleanup] Remove last usages of DirectIngestRegionLockManager (Recidiviz/recidiviz-data#28006)
## Description of the change

The `DirectIngestRegionLockManager` managed two types of locks, one that
was a "process-based" lock that notified other processes that the
/extract_and_merge endpoint was running / writing data to the DB, and
one that was a "resource-based" lock that told other processes that a
certain GCS raw data file was being processed. The former is no longer
actually added by ingest or read by any other process. The latter is
only referenced internally to raw data import to prevent from issues
when Google double-executes Cloud Tasks.

This PR removes the `DirectIngestRegionLockManager` and replaces the raw
data file locking usage with a lock created with a non-custom
`GCSPseudoLockManager`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0c2f53cc0ac851900bf14f838bb6d1fde3fe64cd