[IID Cleanup] Remove last usages of InstanceIngestViewContents (Recidiviz/recidiviz-data#28023)
## Description of the change

Removes the last two references to the legacy
`InstanceIngestViewContents` which was used to query / update the legacy
materialized ingest view results:

1) Deletes the unused `materialize_view_for_args()` function and all
related code. The `IngestViewMaterializerImpl` class is now just a
wrapper around a collection of class functions (more cleanup there to
come).
2) Updates the backend of the Data Freshness admin panel page to just
return `None` as the data freshness value since the values we're
returning right now are already garbage. There are TODOs to scrap the
page entirely and return better freshness info in the ingest instance
status page.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1cb683772b8d78a41c859b028121e68aafb2178a