[Insights] Don't require client names (Recidiviz/recidiviz-data#28111)
## Description of the change

Make `client_name` nullable on the `supervision_clients` and
`supervision_client_events` DB entities.

I originally was going to include a validation for clients that do not
have names, but it looks like there are 229 client ids without names for
supervision clients that have a metric event in the latest year period
in PA (which would be the events that show up on the officer-specific
page). Because of this, I don't think an Insights-specific validation
would be useful now. For now, I think we should merge this change and
then investigate after we get all the data flowing and do UXR.

Query to see client id's with missing names:
```
SELECT
  state_code,
  client_id,
FROM
  `recidiviz-123.outliers_views.supervision_client_events_materialized` e
WHERE
  e.client_name IS NULL
  AND event_date >= DATE_SUB(DATE_TRUNC(CURRENT_DATE('US/Eastern'), MONTH), INTERVAL 1 YEAR)
GROUP BY 1, 2
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 20f19958d261620e35fdbadccb831dd1de921ab0