[IID Cleanup] Removes ingest view raw data CTE materialization option (Recidiviz/recidiviz-data#28084)
## Description of the change

Pre-IID we had a flag you could add to ingest view definitions that
would instruct the query generation to generate separate `CREATE TEMP
TABLE` clauses for each of the raw table CTEs in ingest view queries.
This was a bit of a hack to allow us to get around BQ query complexity
limits we were hitting for some views. When you generate a query with
the extra `CREATE TEMP TABLE` clauses it becomes a "script" and not just
a normal query. Script queries are not allowed from Dataflow pipelines,
so as a requirement of IID rollout, we had to ensure that every ingest
view now is under BQ's (completely obfuscated) complexity threshold and
the query doesn't crash when run in Dataflow.

This PR:
* Removes the `materialize_raw_data_table_views` option entirely from
the ingest view query builder (it was already being ignored when
`using_dataflow=True`
* Removes the default for the `using_dataflow` flag so we can show where
it was implicitly being set to `False` and flip the value to `True`.
Throw if value is ever set to `False`. A follow-up PR will delete the
`using_dataflow` flag entirely.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f612753941bf0855a933d0db8b883abd6f288cb7