[Admin Panel][Insights] Configuration "Promote" functionality - default and production (Recidiviz/recidiviz-data#28026)
## Description of the change

Adds functionality to promote a config to either "default" or to
"production"

I like this implementation better than the previous PR which had buttons
above the table. This seems a little more straightforward about how to
select a config and what the promote options are. Also adds a popup
confirm, and reloads the table after promoting to default so the user
can see the updates.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27556 and Closes Recidiviz/recidiviz-data#27454

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Alexa Batino <37125652+alexabatino@users.noreply.github.com>
GitOrigin-RevId: 4d2bf2d0a269dec53862b43b401200c557ed6c04