[US_AZ] Add datetime_sql_parser to raw data configs (Recidiviz/recidiviz-data#28129)
## Description of the change

In the most recent transfer of data from AZ, they started including
dates in the form "12-29-2024 00:00:00.0000000" when before they had
been in the form "12-29-2024 00:00:00". This change adds a
`datetime_sql_parser` to the raw data configs where this is happening to
remove the .0000000 before attempting to parse the field.

I tested the new configs on all of the files we have received from AZ,
including both the old formatted dates and the new ones. I loaded these
to a sandbox with prefix `elise`.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28115

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b5a69bb89e68de1c35c97b8e72cb7817f661e13c