[Insights] Switch US_IX to US_ID in outliers_config (Recidiviz/recidiviz-data#27986)
## Description of the change

This PR includes changing the OutliersConfig to be mapped to US_ID
instead of US_IX and adding (temporary) logic that returns the config
mapped to ID for IX and uses the BQ-specific list of enabled states to
refer to IX in views, which can be removed after the eventual ATLAS
migration from `US_IX` to `US_ID`.

With these changes, I expect that the product views still use IX as the
state code since the IX data in BQ reflects the ATLAS data–this is why I
crated the `get_outliers_enabled_states_for_biqeuery` method. However,
the views are exported to GCS into a US_ID folder with the export
override and then imported to a US_ID DB instead of the current US_IX
DB, which would become defunct since the `outliers_enabled_states.yaml`
change should initialize a US_ID DB instead of a US_IX.

I'd hold on merging this until _after_ the next batch of ID emails are
sent out, which was scheduled for Thu 3/7, to avoid breaking anything
before then : )

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1bcec02443688d89f77b93bc62698d2d1677b055