[US_ND] Make custodial authority state prison for people on work release (Recidiviz/recidiviz-data#28149)
## Description of the change

Some people on work release appear in state_incarceration_period but not
in location_sessions. We believe this is because the
`StateCustodialAuthority` is `COUNTY` on their incarceration period when
it should be `STATE_PRISON`. This change updates the mapping to assign
folks in those work release facilities specifically to
`StateCustodialAuthority.STATE_PRISON` and leaves the rest of the logic
untouched.

I am running a test ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-03-12_15_06_08-1739725225215275816?project=recidiviz-123)
to ensure that this fixes the problem
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27932

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 45570aa962f67298540b425c94729f375701d5e7