Adds dynamic assignment sessions join (Recidiviz/recidiviz-data#28113)
## Description of the change

Adds params to the function that generates the
`assignments_with_attributes` LookML view that allow for additional
filtering by a dynamic sessionized view of attributes. For now, both
custom metrics views are using `compartment_sub_sessions` and the set of
attributes available in that view.

This will eventually unblock adding filters by workflows location and
workflows caseload type to the custom workflows impact metrics explore.

I opted for a pretty simple but somewhat restrictive set of params here
where there is exactly one dynamic attributes view per lookml view, but
definitely open to supporting 0/2+ attributes views if that feels
appropriate (either in this PR or as a future feature).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related Recidiviz/recidiviz-data#26533

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: f45127fd6051714e5b669cbf37cf25626a2461bf