[US_TN] Restricting incarceration starts to TDOC facility starts (Recidiviz/recidiviz-data#28173)
## Description of the change

A Trusted Tester at MCCX pointed out that someone we surfaced as
eligible for a custody level downgrade had an incorrectly calculated
score, because of [Question
6](https://drive.google.com/file/d/1T3-U1YmMsGngnVC0PEbdkWFhH08fnF1z/view?usp=sharing)
where someone is supposed to score a 0 as a New Admission/Parole
violator *when they enter a TDOC* facility rather than when they enter
incarceration (jail or prison). We were previously not making this
restriction. This PR uses sub-sessions to join onto TDOC facilities and
use those to calculate someone's "critical date".

Validation notes:
- This reduces the number of people eligible for CLD by 120 (890 -> 770)
and I've spot checked the individual in question as well as other
individuals to confirm that the change is due to a change in score for
Q6, which is due to the date someone started at TDOC

```
WITH main AS (
  SELECT  person_external_id
  FROM `workflows_views.resident_record_materialized` r,
  UNNEST(all_eligible_opportunities) AS all_eligible_opportunities
  WHERE state_code = 'US_TN'
  AND all_eligible_opportunities = 'usTnCustodyLevelDowngrade'
), sandbox AS (
    SELECT  person_external_id
  FROM `dsharm20240513_workflows_views.resident_record_materialized` r,
  UNNEST(all_eligible_opportunities) AS all_eligible_opportunities
  WHERE state_code = 'US_TN'
  AND all_eligible_opportunities = 'usTnCustodyLevelDowngrade'
)
select * from main full outer join sandbox using(person_external_id) where sandbox.person_external_id is null
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28165

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c0e08f4a6bdf4945842abf1bfbba14952bf29a8c